requirements.txt
python-dotenv
langchain_community
langchain-text-splitters
langchain_mistralai
langchain-qdrant
pypdf
streamlit
fastapi
uvicorn
websockets